diskq-nonreliable: split push_tail() based on queue segments
This is a refactor, the commit contains equivalent transformations based
on the following considerations:

This commit splits push_tail() into 3 push parts: qout, disk, and qoverflow

- if qout has space in it and the disk is empty, we push into qout
- we try to push into the disk buffer otherwise
- when the disk push fails (full or write error), we push
  into qoverflow
- once we reached the qoverflow area, we continue pushing message there
  instead of disk

The above steps retain FIFO order if messages are moved from qoverflow to
disk, and from disk to qout after each pop(). This is done by _move_disk().

Notes:
Settings local_options.ack_needed to FALSE converts the subsequent
log_msg_ack() call to a NOOP. It is cleaner not to call the ack method.
The same is true for log_msg_ref() and log_msg_unref().

Signed-off-by: László Várady <laszlo.varady@protonmail.com>