logqueue-disk-reliable: add skeleton of qout memory queue
This memory queue will work similar to the one in non-reliable, with the
difference, that we still store the messages on the disk, additionally
to storing it in the memory queue.

The main usage of qout in the reliable disk buffer is to have some msgs
stored in memory, so when we get to the point in the qdisk, where we
also have that msg stored in the memory, we just skip the one on the
disk, and use the one in the memory, saving one message read from the
disk, and one deserialization.

The first difference between qout and qreliable is that we ack the msg,
because we do not want to trigger the flow-control.
The other difference is, that we do not move from qout to qbacklog,
we let the qdisk's backlog handling handle the situation.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>