afsocket: support TLS-passthrough for proxied-tls
This new transport method allows syslog-ng to interpret the (plain text)
V1 proxy protocol header which comes before the encrypted TLS payload.

Signed-off-by: Parrag Szilárd <szilard.parrag@gmail.com>

There is one problem with this change: TLS is now allowed but not
required, so a missing tls() block will be a run-time error, not a
config-parse error. We can fix this later as well, but it also
requires changes in one of our internal modules.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>