nvtable: add "type" member to NVEntry to allow the storing of type information of name-value pairs
Earlier only "indirect" values had type information, but that limits the
application of types related to name-value pairs.

The new field is reusing padding byte that was present in earlier versions
of this struct, thus the size of the struct is not increasing with this
change.

Since NVEntry is directly serialized as a sequence of bytes, a new "flag"
is also added that indicates that the type information is indeed present.

This way, old versions of syslog-ng would store NVEntry as "not-typed"
(it initializes all unknown flags to zero OR our upgrade code in the
deserializer does this), whereas any new version that gets
released with this patch included would set the same flag, indicating
that type information is indeed available and "type" contains a valid value.

Old, previously serialized LogMessage structs are transparently upgraded
in-place by the deserializer: by checking the flag and properly
initializing new fields if they are unset. This way we don't need a
new LogMessage format version.

Old versions of syslog-ng reading the new format would just ignore the new
bits.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>