nvtable-serialize-legacy: only copy flags that are possible in legacy formats
Old versions of syslog-ng didn't guarantee that undefined flags would be
initialized to zero. Set those explicitly and make sure new flags are
properly initialized.

NOTE: I don't understand why we perform an in-place upgrade of the NVTable
instead of simply populating name-value pairs with the existing code. We
have a parallel implementation of NVTable creation in nvtable.c
and nvtable-serialize-legacy.c

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>