dbld: limit setuptools version for kira and devshell image
https://setuptools.readthedocs.io/en/latest/history.html#v58-0-0
setuptools v58.0.0 removed the support for the `use_2to3` keyword.

In kira and devshell image, we are using the hy pip package,
which depends on funcparserlib.

The current funcparserlib version (v0.3.6) uses this unsupported
`use_2to3` keyword, although v1.0.0 is in the making, and it will not
use `use_2to3`.

See:
https://github.com/vlasovskikh/funcparserlib/commit/84f1d35a84f73dd61817f1a32421833118174c28

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>