cfg-parser: add const qualifier to yylloc in LEXER_BINDING macros
Starting with bison 3.8, bison generates the function declarations we
have already done in CFG_PARSER_DECLARE_LEXER_BINDING.

The only difference between the generated one, and our macro is a
const qualifier before CFG_LTYPE *yylloc.

In order to work with both 3.8 and older bison versions, we should
keep our manual function definitions in the code, but it should match
with the one, bison generates.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>