lib/filter/filter_throttle: Add Mutex around RateLimit objects in throttle-filter
Here we wrap the accesses of the ThrottleRateLimit objects that we use to maintain the
ratelimit state in a mutex. This means that we don't get race conditions where two threads
consume the same token from the bucket at the same time, bypassing the ratelimit. By only locking
the specific ThrottleRateLimit that we're operating on for the current message, we don't get _too_
penalised on locking times, assuming that the range of ratelimit keys is decently distributed

Signed-off-by: sinkingpoint <colin@quirl.co.nz>