lib/filter/filter_throttle: Make value optional in throttle filter
This commit updates the throttle filter syntax slightly to allow a call that only
supplies a rate limit. In this instance, we set the key to the empty string for all
logs, such that every log effectively shares the same rate limit. This provides similar functionality
to the existing `throttle` in a destination, but provides a bit more flexibility in choosing what
to do with logs that breach the rate limit

Signed-off-by: sinkingpoint <colin@quirl.co.nz>