actions: disable docker image upload on forked repositories
This change is a result of a long discussion, trying to summarize the
reasons behind it.

There were two main concerns regarding builder images:
Are there anybody who wants to maintain a custom cache for dbld images?
Can we support it without a code change on the forked repository?

The answer for the first question is probably none or very few. Keep in
mind: we still build the docker images in case of dbld code changes, so
one can be sure that nothing breaks. We just skip the upload step.

The answer to the second question is YES. We already do it without this
commit, so this is a step backwards. However using those images in the
GitHub Actions of the forked repository already requires a code change.

example:

1) Remove the if clause from the dbld-images.yml action.
   Or simply revert this commit.
2) Add the env: CONTAINER_REGISTRY: ghcr.io/${{ github.actor }}
   to the selected action. (i.e.: packages.yml)

Signed-off-by: Laszlo Szemere <laszlo.szemere@oneidentity.com>