lib/filter: Add clone virtual method to FilterExprNode
This commit adds in a `clone` optional method to FilterExprNode instances. This method is used
when we clone the surrounding LogFilterPipe instance, and if it exists we call that method to
get a deep clone of the FilterExprNode rather than just incrementing its reference count.

In addition, this commit adds an implementation of this method into FilterThrottle that clones
the rate and key_handle instances into a brand new instance of the Filter. This allows cloned
FilterThrottle instances to maintain their own state, rather than keeping the same ratelimits map
across all instances of the same filter (which would lead to double counting of logs)

Signed-off-by: sinkingpoint <colin@quirl.co.nz>