actions: do not upload dbld images in case of nightly runs
The main purpose of the scheduled job is to detect any problem with the
images as early as possible. Changes in upstream packages can break our
images without a code change on our side.

By not uploading the result, we can preserve the last working version
in the cache. This will give us some time to fix the issue, and will
not block other - non related - PR's from merging.

To avoid code duplication, I extracted the if condition of image upload.

Signed-off-by: Laszlo Szemere <laszlo.szemere@oneidentity.com>