lib/filter: Remove filter_expr_ref in favor of filter_expr_clone
This commit introduces a new method - filter_expr_clone to the filter_expr.h file. In this method, we call
the clone virtual method on FilterExprNode's if it exists, and otherwise fall back to the filter_expr_ref method.
We also swap over all the calls to the _ref variant over to _clone so that clonable Nodes are cloned instead of reference counting

With this change, we also introduce new methods for cloning several FilterExprNode variants including FilterCmp, FilterOp and FilterPipe
so that we can recursively clone those Nodes that reference other FilterExprNodes

Signed-off-by: sinkingpoint <colin@quirl.co.nz>