container_synchronizer: add skeleton of ContainerSynchronizer
This class will be responsible for syncing the contents of the remote
containers back and forth with the local files.

The main idea is, that it will provide the user a directory structure
with the files available in the container, the user can do whatever they
want, then sync back the changes.

As we have multiple directories in a container (nightly, stable, etc...)
it would be good, if this class would only operate on a certain subdir
of the container. This can be set with the sub_dir variable in the ctor.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>