filter/throttle: remove value option
The new template support were added to throttle filter, a modification
only requires users to change value("A") --> template("${A}").

The template is a better option to have, the only downside could be
performance, but it is possible to use a lightweight version of template
called trivial template that is quick.

As all template containing only one NV key is trivial, this should not
cause a regression.

As the value option was never published in a release, it is safe to
remove and only provide template option.

Signed-off-by: Kokan <kokaipeter@gmail.com>