deb_indexer: remove old nightly packages
The algorithm leaves only the packages with the 10 freshest upload date.
It is not the same as only leaving the packages, which have fresher
upload date than the (current date - 10 days). I was thinking about what
happens, if for some reason a platform's packages fail to build nightly
for 10+ days. I think we do not want to remove the old packages in this
case.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>