container_synchronizer: delete snapshots too, when deleting blobs
According to:
https://docs.microsoft.com/en-us/python/api/azure-storage-blob/azure.storage.blob.blobclient?view=azure-python#delete-blob-delete-snapshots-none----kwargs-

> Note that in order to delete a blob, you must delete all of its
> snapshots. You can delete both at the same time with the delete_blob()
> operation.

------------------------------------------------------------------------

Why is it safe to delete snapshots?

Snapshots give us the chance to restore overwritten blobs, but they
do not give us a chance to restore deleted blobs, as when we delete
a blob we must delete all of its snapshots with it.

In order to be able to restore a deleted blob, soft deletion is the way
to go. See:
https://docs.microsoft.com/en-us/azure/storage/blobs/soft-delete-blob-overview
Basically it keeps our deleted files for x days, and we can restore them
if necessary. We set this to 10.

When de we even want to delete a blob / snapshot?
  1. When we processed the incoming container.
  2. When a nightly package in indexed gets "old".
  3. We DO NOT want to delete any file from the stable release's indexed
     container.

What this commit changes in the 3 cases above?
  1. We do not create snapshots from blobs in the incoming container,
     so we did not change anything.
  2. We now are able to remove old packages and their snapshots and it
     is okay to do so, because we still have 10 more days of restoring
     deleted blobs because of the soft deletion.
  3. With the previous commit, we do not give the permission for the
     stable release indexer to delete a blob, so this commit does not
     change anything.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>