json-parser: convert incoming JSON arrays to a syslog-ng list
Earlier, JSON arrays were converted to name value pairs with the array index
enclosed in the key within brackets, for example:

{'array':[1,2,3]}

This was converted to 3 name-value pairs:

  array[0]="1"
  array[1]="2"
  array[2]="3"

This was not documented, its usability was limited and a $(format-json)
couldn't reproduce the same list, even though $(format-json) has supported
syslog-ng lists for a while now.

This patch changes this in an incompatible way, the array becomes a
syslog-ng list that you can use with the $(list-*) family of template
functions, which means that $(format-json) would transparently start working
with this as well.

NOTE: that the list format is only used as the list has only primitive types
(strings, integers, booleans, nulls), if a complex type (array or object)
element is encountered as a list item, then the entire array becomes a
literal() value, which is supported by $(format-json) but the list functions
would not work on it.  It is expected that the schema of these objects is
relatively static, so as long as the configuration uses application specific
configuration elements, this should not cause an issue.

Because of this being an incompatible change, this is considered RFC.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>