json-parser: add support for top-level JSON arrays
Top-level JSON arrays were not supported so far and causes json-parser()
to fail. This patch adds the capability to parse JSON array elements into
$1, $2, $3 ... and so on.

NOTE: this is different to what we do in the embedded array case which are
translated into lists that we can later process with the $(list-*) template
functions. The reasons the two cases are orthogonal is that at the
top-level we don't have a key for the name-value pair which would hold
the array elements.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>