remote_storage_synchronizer: set sub_dir separately
This change seems unnecessary at this point.
It will be needed later on, for the following reason:

Currently we are constructing an `AzureContainerSynchronizer` intance in
`Indexer` and at that point `DebIndexer` already told what the `sub_dir`
is. But we are plannig to move the construction of any
`RemoteStorageSynchronizer` outside Indexer and `DebIndexer`, and it
will just be passed to `DebIndexer`. As `RemoteStorageSynchronizer` is
not constructed in `Indexer, `Indexer` has to manually set the `sub_dir`
for the `RemoteStorageSynchronizer` instance it received.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>