qdisk: fix layering violation in serialization code
QDisk is responsible for storing any kind of binary data in a file.

Currently the serialization and deserialization functions operate on
LogMessage-s not on a general binary data, which is a layering
violation. The knowledge of how to serialize a LogMessage is the
responsibility of LogQueueDisk.

This commit refactors the code around serialization in QDisk and
LogQueueDisk so that the above mentioned responsibilities are met.

Also, QDisk's serialization can be a class method.

Note:
The deserialization functions have the same problem. Their refactor is
in the next commit.

Signed-off-by: Balázs Barkó <Balazs.Barko@oneidentity.com>
Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>