github-actions: migrate ARM build from Travis CI to github-actions
As we focus more on GitHub Actions, for testing builds on ARM
we could use a github-action that's using Docker + QEMU on Linux,
instead of Travis CI.
See: github-action: https://github.com/uraimo/run-on-arch-action

This patch migrates the current Travis config to the run-on-arch github-action.
Following patches make it functional (some packages that are pre-installed in Travis has to be installed),
and extend it.

Notes:
 - It's not a native ARM runner though, it uses QEMU
 - Build time is slow: even after putting dependency installation (including locally-built ones),
   syslog-ng build is still around 27 minutes (Github runners have 2-core CPUs, but it was slow on my 4-core i7 CPU too)
 - It requires a docker image to be built first, in order to run.
   We can cache the docker images in the Github Packages (similarly to dbld images)

Alternative github-action:
 - https://github.com/pguyot/arm-runner-action
    Similarly it can run ARM-based binaries with QEMU.
    Not docker-based, it uses a chroot environment and an image of the given distribution.
    It also creates an image that can be used on Raspberry PI.
    Currently it is limited to Raspberry PI.

Signed-off-by: Gabor Nagy <gabor.nagy@oneidentity.com>