logmsg: rename LM_VT_LITERAL to LM_VT_JSON
We have the literal() type-hint which was used to indicate for
$(format-json) to include a value as is.  However with name-value associated
typing I feel it is better described that it is a JSON snippet rather than
something "generic" literal.

The literal() type-hint is still accepted for compatibility however one now
can use json() as a type-hint. And I think it is also more readable.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>