cdn: add factory-like function to help creation from config
These functions will help us instantiate different implementations of
the `CDN` interface.

I think it is a bit of a layering violation, that the `CDN` class markes
these functions abstract on its own, maybe this should be done with a
`ConfigElementFactory` object with composition but I did not want to
overcomplicate this part. Maybe in the future we could think about it.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>