control: eliminate mainloop dependency in ControlCommandThread
While running the test programs we are not able to execute control commands
in separate tests. For this reason, the ControlCommandThread had a
conditional to see if mainloop is running.

This was a layering violation only needed for test purposes so this patch
eliminates this by providing a _run_command() method implementation in
the test related dummy class.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>