cfg-grammar: use LL_PLUGIN token types for plugin insertion points
This patch changes how plugins are recognized in the configuration file.  To
make the change easier to grasp, here a short summary how it works today:

Example configuration:

source whatever { internal(); plugin(); };

Here, the source driver internal() is built into the syslog-ng core,
plugin() is provided by a loadable module.  When syslog-ng parses the source
statement, it would match ';' separated driver references.

For built-in drivers, we have dedicated keywords, for example the "internal"
word is lexed as KW_INTERNAL, which then causes the internal() driver
related rule in the grammar to match.

Plugins on the other had are not built into the core grammar, thus their
keyword is not recognized by the lexer.  In fact the external module may not
even be loaded yet, when we process the configuration that has a reference
to a plugin.

What happens is that the lexer returns LL_IDENTIFIER, a token type that is
reserved for any unqualified word-like input.  This is a pretty generic
token, for example the name of the source block "whatever" is also returned
as LL_IDENTIFIER.

The grammar has a rule for recognizing plugin references, e.g. something like this:

source_plugin: LL_IDENTIFIER { ... look up, load and instantiate the plugin ... }

This was working fine until now, however I want to introduce rules in the
filter context that would also use LL_IDENTIFIER tokens as legitimate part
of rules, not used from plugin instantiation. But this conflicts with this use.

With all this said, this change will cause the lexer to return LL_PLUGIN
token for any discovered plugin names, allowing me to resolve this grammar conflict.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>