java: configuration replace
The log4j allows us to replace the configurator object with our own,
allowing the user to programmaticly configure log4j. In our case we want
to replace all appender with our implementation, this is perfectly align
with that.

Note: with this user cannot change how log4j is logging.

Signed-off-by: Kokan <kokaipeter@gmail.com>