pragma-grammar: deprecate module-loading related defines
We are removing support for changing these values from within the
configuration, as it complicates the startup sequence a great deal and
that I fail to see valid use-cases that warrant these complications.

Using my google-fu I made an effort to searching any examples of the use
of these defines and apart from the documentation I couldn't find any.

This patch adds explicit deprecations for these values.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>