logmatcher: set num_matches in the LogMessage
Previously the number of matches were lazily tracked by log_msg_set_match()
calls, e.g. once a new match was set outside of the existing range,
the size of the range was bumped.

This is not true however in the case when a subsequent match is performed,
which might generate fewer matches.

This patch change LogMatcher to always update num_matches, this way
it represents the result of the _last_ matching operation.

The old behavior where we update num_matches remains, as the user is
also permitted to set() $1, $2, etc. In that case num_matches attempts to
follow the changes of the match values.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>