nvtable: get rid off nv_table_get_value_if_set() variant
The only difference between nv_table_get_value() and
nv_table_get_value_if_set() is that nv_table_get_value() never returned
NULL, but returned a string with zero length instead.

A similar logic is implemented in the LogMessage layer, so get rid off
this duplication by:
  * nv_table_get_value() WILL return NULL for an handle that has no value
  * log_msg_get_value() and friends would handle this NULL and return the
    zero-length string

This reduces some logic duplication as well as centralizes a place where
I can set the type for this zero-length string. This is to be
done in a subsequent commit.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>