logqueue-disk: store compaction option in LogQueueDisk
Currently, there is a race condition in the disk-buffer's code:
  * There is serialization happening without holding the lock.
  * Meanwhile, another thread could be running
    log_queue_disk_restart_corrupted().

In theory the two should not cause problems to each other, as while we
serialize, we do not operate on the disk-buffer file, so it can be
replaced.

However, the serialization code currently uses QDisk->options, which
gets replaced in log_queue_disk_restart_corrupted().
If we remove the usage of QDisk->options while serializing, we have
solved the problem.

So why do we use QDisk->options? Because we need the compaction field
from it. However the compaction flag is related to LogMessage, so it
has nothing to do inside QDisk's options. The options of LogQueueDisk
and QDisk are mixed into one Options class, which is not great.

In an ideal case, QDisk should have a separate options class, which
should only contain disk_buf_size, mem_buf_size, etc...
Similarly, LogQueueDisk should have a separate options class, which
should only contain compaction, reliable, etc...

To do that we would need a relatively big refactor, which we do not have
the time to do. So as an uglier solution I only added a new compaction
field to LogQueueDisk, alongside with a TODO comment.

Note:
QDisk->options->compaction is used in another function: _save_queue()
It is handled in the next commit.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>