qdisk: remove usage of compaction field
Any kind of LogMessage related code in QDisk's code is layering
violation, just as this one in _save_queue().
Getting rid of the LogMessage dependency in QDisk is a bigger bite, so
log_msg_serialize() stays for now.

However we can get rid of the usage of the compaction field, which has
its pros and cons:

Pros:
  * We remove a bit of the layering violation.
  * We remove a possible race condition.
    (Deducing that whether it is or is not accessed while another thread
    is running log_queue_disk_restart_corrupted() is hard, and can
    change easily.)

Cons:
  * We might store larger messages than it would be necessary. IMO this
    is not a big problem, as we are already writing over the desired
    file size, also not a lot of messages are stored in the memory
    queues, so it does not make a huge difference.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>