cfg-grammar: fix references to last_template and last_filter_expr in case of recursive parsing
It is a general practice to use global variables to instantiate various
objects while the grammar is running.  For instance last_filter_expr is a
variable that holds the filter expression that is being parsed.

The issue with this practice is that sometimes our grammar recurses into
itself, for instance a template function may have a filter argument, e.g.
$(context-lookup). And a filter expression may reference a template.

In these cases, the recursing grammar will overwrite the
last_template/last_filter_expr variables causing

a) some of the options being applied to the wrong instance (e.g.  when we
overwrite the value and have some more to parse)

b) using the wrong instance referenced, as we use "$$ = last_template" style
returns.

This patch intends to address these cases by using rule-local variables to
represent the value in question.

The globals are not yet eliminated by this patch, so I am not entirely sure
this is a complete solution.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>