qdisk: do not exceed max size when the last message fits properly
qdisk_is_space_avail() was not consistent with the writing logic of QDisk,
so it was possible to exceed the maximum size of the buffer in case the
last message fitted completely.

This was possible only when there was enough space at the beginning of the
disk buffer. In that case, instead of writing to the beginning, syslog-ng
wrote to the end of the file, exceeding the maximum buffer size for no
good reason.

This has been fixed without bumping the header version, so in extremely
rare cases, a single message may be lost during upgrade.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>