qdisk: fix a rare disk buffer corruption issue
A disk-buffer is considered empty when backlog_head (and read_head)
points to write_head. The same could have been done when the
buffer is full, but currently we don't accept that, because we can't use
`length` to distinguish between these two states.

Due to the above, when a disk-buffer is filled completely, write_head will
point to the end of the disk-buffer file as backlog_head prevents the
wrapping logic to do its job.

In that specific case, if the disk-buffer then gets emptied, the read and
the backlog heads will point to the end of the buffer as well.
An additional write  will wrap the write head and allow the read head
to read into an invalid position, which ends up detected as an invalid
disk-buffer file.

This has been fixed by adding a preliminary check to each disk-buffer
operation (qdisk_skip_record() is used for moving backlog_head).

Note: we should probably allow "read_head == backlog_head == write_head"
situations when the buffer is full in the future in order to be completely
consistent with the size of the buffer (if we don't allow that, we can
fill a 400 bytes long disk-buffer with two 200-byte messages, but we can
NOT pop one of them and then push it back, because we require a single byte
empty space between backlog_head and write_head).

Signed-off-by: László Várady <laszlo.varady@protonmail.com>