mainloop-threaded-worker: split run into thread_init/run/thread_deinit
Instead of pushing the responsibility of calling
main_loop_threaded_worker_signal_startup_finished(), let's split the run()
method, so that we can indicate the startup result ourselves
based on the return value of the thread_init() function.

NOTE: this patch introduces a naming confusion as
MainLoopThreadedWorker->thread_init and LogThreadedDestDriver->thread_init
has slightly different purpose, so I intend to rename the latter in
a subsequent commit.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>