logthrsourcedrv: move LogThreadedSourceDriver related methods to the driver
There are some methods that are indeed worker specific (in their name and
their function), however they take a LogThreadedSourceDriver * as their
self pointer.

Move these "foreign" methods to their home and get rid off the setters that
masked that they were somewhere else.

Also the setters were needed earlier as the Worker class was private.

A subsequent commit would fill these function pointers in the constructors
as it should.

In a way these methods a quite akin to the compat layer in LogThreadedDestDriver.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>