disk-buffer: fix "memory leak" and inconsistent buffer handling
The reliable disk buffer maintains in-memory copies of the "beginning" and
the "end" of the disk buffer for performance and flow-control purposes.

These memory buffers are not necessarily continuous slices of the disk
buffer file, so they are emptied by a logic that queries the reader head of
QDisk.

Currently, it is not allowed for the reader and writer heads to overlap
when the buffer is full, so it is possible that the read head points to
the very end of the file, but the next read will be done from the beginning
of the file.

This corner case has to be taken into account when comparing the current
position with in-memory copies.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>