template: change log_template_set_type_hint(NULL) to that it disables type-casting
Earlier, passing NULL as a type hint to log_template_set_type_hint() meant
to cast the result to string.  This behavior is changing with this patch,
such cases mean that casting is going to be disabled.  This casting could
have been set by log_template_new() due to compatibility with 3.x.  This
compatibility can be disabled in the case of value-pairs (which does this),
as it performs its own compatibility functions.

The case of handling type_hint == NULL is going to be handled in the
LogTemplate layer, so handling NULL values in type_hint_parse() is not
needed anymore.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>