template: add support for type propagation in templates
Template strings sometimes contain relatively trivial expressions (e.g.
a single ${value}), in which case we can propagate the type information
of the value as the result of the template expression.

This patch adds LogTemplate methods that allow returning of the
type information and implements type propagation for expressions that
only have a single value reference.

Compatibility: we work as if an explicit string() type-hint was specified in
3.x mode, effectively casting our output to be a string.  This is lifted in
4.x mode.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>