dbld: use dedicated RPM mirror for fedora-33
https://mirrors.fedoraproject.org/ often returns 503 error code, when
we are trying to pull the metalink.

We can work this around by using a dedicated mirror, and not trying to
look for mirrors via the metalink file at all.

The actual server was chosen from this list:
https://admin.fedoraproject.org/mirrormanager/mirrors/Fedora/33/x86_64

I tried to choose a server which is EU based, and has large bandwidth.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>