mainloop-worker: delay calling callbacks scheduled by main_loop_worker_sync_call()
The mainloop performs config reload in multiple stages. These steps need
to be synchronized with the log pipeline running within syslog-ng.

Whenever we want to swap out the configuration, we need to stop
all workers first and wait for them to exit. This is the intent of
main_loop_worker_sync_call().

If there are no workers, the callback is invoked immediately. If there
are workers still running, it requests workers to exit and
schedules a callback that is going to be called when the number of
workers reaches zero.

Without this patch, these callbacks are invoked right from within the
worker completion callback. At that point the worker is still running,
as it is waiting for the result of its completion callback.

There's a deadlock scenario, introduced by the latest mainloop-threaded-worker
changes:
  * [M] the mainloop is trying to reload/exit, calls main_loop_worker_sync_call()
    with a callback to commence with reload/exit
  * [W] the worker thread of the internal() source is running and that is
    the one that is exiting last
  * [W] within its thread_deinit() function the worker thread calls the
    completion callback using main_loop_call()
    and blocks waiting for the result.
  * [M] the mainloop gets the completion, calls the scheduled callbacks to
    commence with the reload
  * [M] as a part of reload, LogPipe instances are freed,
  * [M] afinter_source_free() is called, which calls
    main_loop_threaded_worker_clear()
  * [M] a recent addition to main_loop_threaded_worker_clear() is to join
    the worker thread before cleaning up resources. We block here.
  * [W] that thread is at this point is waiting for its main_loop_call() to complete
  * [M] this never happens as the mainloop is waiting in g_thread_join()

The solution was simpler than I thought: we delay
_invoke_sync_call_actions() with one ivykis iteration. This allows
the worker's main_loop_call() to complete. Thus it is able to exit, so that
g_thread_join() will not block. Deadlock solved.

I've found that we already do some of the house-keeping in the
completion callback as an iv_task, so I simply moved the
_invoke_sync_call_actions() to the this iv_task.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>