logproto: move split_buffer() to LogProtoBufferedServer
split_buffer() moves data from the middle/end of the buffer to the
beginning, which is the responsibility of LogProtoBufferedServer.

Currently, the split logic is not called when the end of the buffer is
reached, it is called on specific LogProtoTextServer events, so the method
is currently "protected" instead of "private".

Signed-off-by: László Várady <laszlo.varady@protonmail.com>