mainloop-io-worker: generalize job argument
The I/O worker job got a GIOCondition argument but it was not
even used in a couple of cases. This patch changes the argument type to
gpointer so that "any" job argument becomes possible. Previous uses where
GIOCondition was actually used now use GINT_TO_POINTER to pass the same
value.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>