loggen: handle embedded NULs in a message
When reading the messages from a file, handle embedded NULs in case we are
not parsing those messages.

This assumes that fgets() actually returns all characters up to the NL
character, we just need to determine how many characters we have read. We
do that by zeroing out the buffer and then looking for the last non-zero
character.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>