mainloop-worker: increase the maximum number of threads to 256
This patch removes the 64 thread limitation from _allocate_thread_id and
increases the current maximum to 256.

Allocations that depend on per-thread state currently only use the
actual number of threads allowed by the --worker-threads command
line option (which defaults to the number of cores in the system),
so increasing this limit does not cause an immediate increase in
memory usage.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>