logmsg: move log_msg_write_protect() out of log_msg_alloc_*queue_node()
A message needs to be write protected as soon as it is kept for a period
of time (e.g. when other consumers of the same LogMessage may change
name/value pairs, etc)

Traditionally, whenever it enters a LogQueue, it needs to be protected
so once delivered it remains the same.

This write protection was in some cases enabled explicitly (e.g. in patterndb)
sometimes implicitly by the call to log_msg_alloc_queue_node(). This
asymmetry is removed with this patch, log_msg_write_protect() always
needs to be called explicitly, even in the LogQueue case.

This makes it possible to reuse the LogMessageQueueNode mechanism for
other non-queue purposes (e.g. LogScheduler).

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>