logqueue-fifo: save the actual number of threads at constructor time
Instead of relying on a single global log_queue_max_threads variable,
save it at constructor time.

Also, handle the case where we get a thread_id that is larger than this
saved value, instead of failing an assert.

This patch is a preparatory step to allow threaded sources to use the
scalable input path for a LogQueue, because if we want to support that
we need to make the thread_id space configuration dependent.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>