logthrdestdrv: don't set sequenceId for non-local messages
The SDATA meta.sequenceId is the concept behind our SEQNUM macro, e.g.  it
is a destination specific sequence number of local messages being sent.

But this sequenceId only needs to be assigned to messages where we are the
originator, as described here:

    7.3.1.  sequenceId

       The "sequenceId" parameter tracks the sequence in which the
       originator submits messages to the syslog transport for sending.  It
       is an integer that MUST be set to 1 when the syslog function is
       started and MUST be increased with every message up to a maximum
       value of 2147483647.  If that value is reached, the next message MUST
       be sent with a sequenceId of 1.

NOTE the word "originator" in the first sentence.  Non-local messages don't
have a locally generated SEQNUM unless the original message (coming from its
originator) had one.

There are two ways sequenceId are associated with incoming messages: the
structured data meta.sequenceId, or in the case of Cisco messages the
sequence number prepended to each messages.

Now, LogThreadedDestDriver based plugins don't ever implement RFC5424, but
$SEQNUM as macro is available even in these cases, and up to this patch, we
generated a sequenceId for all outgoing messages, local and non-local alike.

Our $SDATA macro then automatically patches in the sequence number if one is
specified by the driver which was the case.  This meant that if you are
using the $SDATA macro in the context of http() for instance, your SDATA
will magically include an extra meta.sequenceId element, which is not what
you want.  This behavior is fine when _we_ are the originator of said
message (as in that case, syslog-ng itself is free to add such an meta
element), but can be confusing and to be avoided when we are forwarding
incoming logs.

This patch changes LogThreadedDestDriver behavior to match that of LogWriter
instances.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>